Add end-to-end tests (#9)
Creates five jobs in two stages, the first four run in parallel and build
the seperate microservices. The last job pulls all microservices,
zookeeper, kafka and nginx and runs end-to-end tests on that setup.

Docker images are pushed to the wdmk(Web-scale Data Management Kafka)
namespace on Docker Hub. Images with tags are tagged and the master
branch produces the latest images. This might become aws or we push to
both.

This means that you can do docker-compose pull (after this is merged)
and docker-compose up to get up and running. Moving to Kube should be
straightforward at this point.

I altered the gradle setup a bit, because spring boot and rest assured
both depend on different versions of gson. This gave problems. I am not
sure if this is the best way to do this or if there is a way to have a
seperate project inside a project in another way. It is important that
the end to end tests are not called by accident, because they will l